domain: move to a sub-package, so that the root package can be used by the public API
