+++ domain/kube-score.go
@@ -1,4 +1,4 @@
-package kube_score
+package domain
 
 import (
 	appsv1 "k8s.io/api/apps/v1"
